Add UnifiedTransformer README
