+++ modules/text/text_generation/unified_transformer-12L-cn-luge/requirements.txt
@@ -1,0 +1,1 @@
+sentencepiece
