+++ modules/text/text_generation/unified_transformer-12L-cn/requirements.txt
@@ -1,0 +1,1 @@
+sentencepiece
